remove redundant cast to int

+++ src/nvim/lua/xdiff.c
@@ -68,7 +68,7 @@ static void get_linematch_results(lua_State *lstate, mmfile_t *ma, mmfile_t *mb,
 {
   // get the pointer to char of the start of the diff to pass it to linematch algorithm
   char *diff_begin[2] = { ma->ptr, mb->ptr };
-  int diff_length[2] = { (int)count_a, (int)count_b };
+  int diff_length[2] = { count_a, count_b };
 
   fastforward_buf_to_lnum((const char **)(&diff_begin[0]), (linenr_T)start_a + 1);
   fastforward_buf_to_lnum((const char **)(&diff_begin[1]), (linenr_T)start_b + 1);
